pyproject.toml
[project]
name = "transcript-video-to-text-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.5",
    "pathlib>=1.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "youtube-transcript-api>=1.2.3",
]
